+++ project-base/src/Shopsys/ShopBundle/Component/Error/Reporting/ErrorReportingFromLastHourCronModule.php
@@ -4,7 +4,7 @@ namespace Shopsys\ShopBundle\Component\Error\Reporting;
 
 use DateTime;
 use Shopsys\Environment;
-use Shopsys\ShopBundle\Component\Cron\SimpleCronModuleInterface;
+use Shopsys\Plugin\Cron\SimpleCronModuleInterface;
 use Shopsys\ShopBundle\Component\Setting\Setting;
 use Shopsys\ShopBundle\Model\Mail\MailerService;
 use Shopsys\ShopBundle\Model\Mail\MessageData;