+++ project-base/tests/ShopBundle/Unit/Component/Cron/CronServiceTest.php
@@ -3,9 +3,9 @@
 namespace Tests\ShopBundle\Unit\Component\Cron;
 
 use PHPUnit_Framework_TestCase;
+use Shopsys\Plugin\Cron\SimpleCronModuleInterface;
 use Shopsys\ShopBundle\Component\Cron\Config\CronModuleConfig;
 use Shopsys\ShopBundle\Component\Cron\CronService;
-use Shopsys\ShopBundle\Component\Cron\SimpleCronModuleInterface;
 
 class CronServiceTest extends PHPUnit_Framework_TestCase
 {